Add `classifai_ocr_text_post_args` filter
This allows for a number of customizations, such as being able to store the text results in a different field in case you use post_content extensively in the editorial workflow already, or set other meta, or update the alt text based on the text results, and so forth.

Also adds the full $scan data as context for the `classfai_ocr_text` fitler.